refactor:集計目標値の更新フォームをコンポーネント化 close #68
